assets folder initial restructure
