+++ examples/assets/envmaps/Mono_Lake_B.hdr
