+++ examples/assets/envmaps/garage.hdr
