+++ examples/assets/envmaps/garage.txt
