+++ examples/assets/materials/plastic-glow.material/plastic-glow_basecolor.png
