+++ examples/assets/materials/plastic-glow.material/plastic-glow_emissive.png
