+++ examples/assets/materials/plastic-glow.material/plastic-glow_n.png
