+++ examples/assets/materials/plastic-green.material/plastic-green_basecolor.png
