+++ examples/assets/materials/plastic-green.material/plastic-green_metallic.png
