+++ examples/assets/materials/plastic-green.material/plastic-green_n.png
