+++ examples/assets/materials/plastic-green.material/plastic-green_roughness.png
