+++ examples/assets/materials/plastic-red.material/plastic-red_n.png
