+++ examples/assets/models/AnimatedMorphCube.bin
