+++ examples/assets/models/AnimatedMorphCube.gltf
