+++ examples/assets/models/CesiumMan.gltf
