+++ examples/assets/models/CesiumMan.jpg
