+++ examples/assets/models/CesiumMan0.bin
