+++ examples/assets/models/stanford-dragon.js
