+++ examples/assets/textures/checkerboard.png
