+++ examples/assets/textures/rainbow.jpg
