+++ examples/assets/textures/roughness-test.png
