+++ examples/assets/textures/uv-wide.png
